# openapi.json
{
  "components": {
    "schemas": {
...
      "CreateArticleDto": {
        "properties": {
          "customPrompt": {
            "maxLength": 500,
            "type": "string"
          },
          "feedProfile": {
            "enum": [
              "default",
              "technology",
              "politics",
              "business",
              "health",
              "science",
              "brasil",
              "teclas"
            ],
            "type": "string"
          },
          "generateAudio": {
            "type": "boolean"
          },
          "url": {
            "format": "uri",
            "type": "string"
          }
        },
        "required": [
          "url",
          "feedProfile"
        ],
        "type": "object"
      },
...
      "GenerateUploadUrlDto": {
        "properties": {
          "articleFileName": {
            "type": "string"
          },
          "contentType": {
            "type": "string"
          },
          "fileSize": {
            "type": "number"
          },
          "s3Bucket": {
            "type": "string"
          }
        },
        "required": [
          "articleFileName"
        ],
        "type": "object"
      },
...
      "ProcessMarkdownArticleDto": {
        "properties": {
          "customPrompt": {
            "maxLength": 500,
            "type": "string"
          },
          "feedProfile": {
            "enum": [
              "default",
              "technology",
              "politics",
              "business",
              "health",
              "science",
              "brasil",
              "teclas"
            ],
            "type": "string"
          },
          "generateAudio": {
            "type": "boolean"
          },
          "s3Bucket": {
            "type": "string"
          },
          "s3Key": {
            "type": "string"
          }
        },
        "required": [
          "s3Key",
          "feedProfile"
        ],
        "type": "object"
      },
...
      "UnauthorizedResponseDto": {
        "properties": {
          "message": {
            "example": "Unauthorized",
            "type": "string"
          },
          "statusCode": {
            "example": 401,
            "type": "number"
          }
        },
        "required": [
          "statusCode",
          "message"
        ],
        "type": "object"
      },
...
      "ValidationErrorResponseDto": {
        "properties": {
          "error": {
            "example": "Bad Request",
            "type": "string"
          },
          "message": {
            "example": [
              "source_id must be a valid UUID"
            ],
            "items": {
              "type": "string"
            },
            "type": "array"
          },
          "statusCode": {
            "example": 400,
            "type": "number"
          }
        },
        "required": [
          "statusCode",
          "message",
          "error"
        ],
        "type": "object"
      },
...
    }
  },
  "info": {
    "contact": {},
    "description": "",
    "title": "Meridiano API",
    "version": "1.0"
  },
  "openapi": "3.0.0",
  "paths": {
    "/api/articles": {
      "get": {
        "operationId": "ArticlesController_listArticles",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Paginated list of articles"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UnauthorizedResponseDto"
                }
              }
            },
            "description": "Missing or invalid bearer token"
          }
        },
        "summary": "List the authenticated user's articles",
        "tags": [
          "Articles"
        ]
      },
      "post": {
        "operationId": "ArticlesController_create",
        "parameters": [],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateArticleDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Article scraped and queued for processing"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationErrorResponseDto"
                }
              }
            },
            "description": "Request failed class-validator validation"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UnauthorizedResponseDto"
                }
              }
            },
            "description": "Missing or invalid bearer token"
          }
        },
        "summary": "Scrape a URL and queue the resulting article for processing",
        "tags": [
          "Articles"
        ]
      }
    },
...
    "/api/articles/jobs/{jobId}": {
      "get": {
        "operationId": "ArticlesController_getJobStatus",
        "parameters": [
          {
            "in": "path",
            "name": "jobId",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Job status retrieved"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UnauthorizedResponseDto"
                }
              }
            },
            "description": "Missing or invalid bearer token"
          },
          "404": {
            "description": "Job not found"
          }
        },
        "summary": "Get the status of an article processing job",
        "tags": [
          "Articles"
        ]
      }
    },
    "/api/articles/markdown": {
      "post": {
        "operationId": "ArticlesController_processMarkdownArticle",
        "parameters": [],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProcessMarkdownArticleDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Markdown article queued for processing"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationErrorResponseDto"
                }
              }
            },
            "description": "Request failed class-validator validation"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UnauthorizedResponseDto"
                }
              }
            },
            "description": "Missing or invalid bearer token"
          }
        },
        "summary": "Queue a markdown article stored in S3 for processing",
        "tags": [
          "Articles"
        ]
      }
    },
    "/api/articles/upload-url": {
      "post": {
        "operationId": "ArticlesController_generateUploadUrl",
        "parameters": [],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GenerateUploadUrlDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Presigned upload URL generated"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationErrorResponseDto"
                }
              }
            },
            "description": "Request failed class-validator validation"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UnauthorizedResponseDto"
                }
              }
            },
            "description": "Missing or invalid bearer token"
          }
        },
        "summary": "Generate a presigned S3 upload URL for an article file",
        "tags": [
          "Articles"
        ]
      }
    },
    "/api/articles/{id}": {
      "delete": {
        "operationId": "ArticlesController_deleteArticle",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Article deleted"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UnauthorizedResponseDto"
                }
              }
            },
            "description": "Missing or invalid bearer token"
          }
        },
        "summary": "Delete an article by id",
        "tags": [
          "Articles"
        ]
      },
      "get": {
        "operationId": "ArticlesController_getArticle",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "includeAudio",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Article retrieved"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UnauthorizedResponseDto"
                }
              }
            },
            "description": "Missing or invalid bearer token"
          },
          "404": {
            "description": "Article not found"
          }
        },
        "summary": "Get an article by id",
        "tags": [
          "Articles"
        ]
      }
    },
    "/api/articles/{id}/audio": {
      "post": {
        "operationId": "ArticlesController_generateAudio",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "202": {
            "description": "Audio generation job accepted"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UnauthorizedResponseDto"
                }
              }
            },
            "description": "Missing or invalid bearer token"
          }
        },
        "summary": "Generate audio for an article",
        "tags": [
          "Articles"
        ]
      }
    },
    "/api/articles/{id}/audio/status/{jobId}": {
      "get": {
        "operationId": "ArticlesController_getAudioJobStatus",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "jobId",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Audio job status retrieved"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UnauthorizedResponseDto"
                }
              }
            },
            "description": "Missing or invalid bearer token"
          },
          "404": {
            "description": "Audio job not found"
          }
        },
        "summary": "Get the status of an article audio generation job",
        "tags": [
          "Articles"
        ]
      }
    },
...
  },
  "servers": [],
...
}